Allow peers to signal Phase hold
This commit allows peers of the `Phase` to signal to the signaler that they have
not yet completed their work. We achieve this by keeping a tally of the total
number of peers that have called `register` -- a new function -- and require
that each peer maintain a `Token` which on drop adds permits to a new
semaphore. On the signaler side if the number of peers equals the number of
permits in the ack'ing semaphore all registered peers have exited.

REF SMPTNG-356

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>