Allow for dynamic block sizes in block construction
This commit improves the 'pack' of our block construction by allow for dynamic
block sizes based around the user-provided block sizes. The key insight here is
that users don't _really_ care about the block sizes and in practice they are
not obeyed: serialized blocks vary depending on the format. They're an
abstraction leak. We therefore dynamically build blocks, keeping track of the
minimum floor for a given serialization format and searching both 20% more and
less on either size of the user-configured blocks.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>