Remove optional return of construct_block
In practice we treated any non-Some return from the function `construct_block`
as a signal that the block construction needed to be retried: the setup was not
satisfiable. The original motivation for this return type was based on a slow
drift of my thinking and that drift is now complete.

Important to tidy up #886

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>