Remove streaming cache method
This commit removes the streaming cache method in favor of the line of work
being pursued in #886. We introduced streaming to work around the inability of
lading to pack its cache space but the downside of streaming is the race it
requires between the generator -- fast -- and the payload creator itself --
slow. This race is always eventually lost by the payload creator and so
streaming can never sustain its throughput goals, except at a very low level.

Sharp edge, that.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>