Experiment with lading telemetry via UDS
We have a need for some lading installations to punch telemetry from inside a
isolated VM or container where only a FS mount is available. This commit
introduces a proof of concept -- in a rough state -- that shows it's possible
for us to create a gauge, counter telemetry stream with registered contexts
to avoid sending the same strings repeatedly over the socket with minimal
coordination between peers.

It's clear some sort of abstraction is still missing here, but it's a start.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>