Split lading::signals::Phase into lading_signals
We know that in practice we have a problem with our Phase mechanism. It had
become overburdened with the sense of multiple transitions being possible and it
was impossible to test. This commit moves the Phase mechanism to its own crate
and our intention is to loom test it. To achieve this the internal complexity
has been significantly reduced, changing the API slightly to rely on ownership,
drops to ensure -- hopefully -- that the necessary semantics of a single signal
transition is maintained.

Loom tests will help us determine if this is true.

REF SMPTNG-455

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>