Split lading::signals::Phase into lading_signals, loom test it
This commit moves the Phase mechanism to its own signals crate and introduces
loom tests for the Phase structure. We believe there is an overflow bug present
in this structure but we are unsure of how to reproduce. This commit introduces
three loom tests, one of which fails, so we hope to have a reproduction case.

REF SMPTNG-455

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>