Don't instrument `Watcher::recv()` (#970)
### What does this PR do?

Optional PR. This may work around failing UDS experiments.

### Motivation

`Watcher::recv()` is cancelled and called again on every request sent by the unix datagram generator and likely others. This is incorrect for the new signaling mechanism and should be adjusted. This PR will eliminate a source of log spam that may be causing experiments to fail to complete.