Expand logrotate metrics' labels (#1027)
### What does this PR do?

I'm seeing in-product that the logrotate file generator bytes_written is
pretty well close to double what it should be. This is concerning and it's
unclear if this is _actually_ a representation of underlying reality or a
consequence of labels being wonky. I'm adding a child_idx label to each
Child to distinguish. Ideally each of these will post 10MiB/s and not 20MiB/s.