Begin work on a logrotate filesystem
This commit starts a series of work that is intended to end in our
first user-defined generator, a FUSE version of our logrotate
file-gen. Although this UDG is bundled in-project I'll puruse an
independent interface and build it in a separate container, eventually,
adding the mechanism in lading for a UDG to be run.

I may eventually migrate `logrotate_rs` into a separate crate. For now
it was convenient to build it in teh `lading` crate.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>