cargo-deny: remove obsolete config fields
The `cargo-deny` crate recently obsolesced multiple fields in its
configuration file schema and use of these fields will now trigger
errors. This commit removes the obsolete config fields and adds
RUSTSEC-2024-370 to the list of ignored advisories for the sake of
preserving the existing state of the project as much as possible while
updating `cargo-deny`'s configuration. Further remediation is deferred
to future pull requests.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>