Begin work on a logrotate filesystem (#1034)
### What does this PR do?

This commit starts a series of work that is intended to end in our
first user-defined generator, a FUSE version of our logrotate
file-gen. Although this UDG is bundled in-project I'll puruse an
independent interface and build it in a separate container, eventually,
adding the mechanism in lading for a UDG to be run.

### Additional Notes

I may eventually migrate `logrotate_rs` into a separate crate. For now
it was convenient to build it in the `lading` crate.