Avoid some allocations, hashes that aren't needed
This commit removes allocations in the FUSE component that don't
need to exist and adjust the interior hashmap to use FxHashMap,
avoiding a secure hash in favor of a fast one. Also, directories
now keep their children in a BTreeSet so that readdir is always
in the same order when called.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>