Add load profile to logrotate FS
We have a need for some experiments to ramp their load up over time.
Current configuration is constant over the lifetime of the experiment,
this change introduces linear growth in addition to constant.

A new release version will be required as this is a breaking change to
logrotate_fs configuration.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>