Attempt generational storage of capture metrics
This commit attempts the generational storage of metrics in the capture
recorder. I am patterning this work from metrics-prometheus-exporter and
I think I understand how this is wired up?

One thing of note gauges are now ints and not f64s although the interface
still requires f64 on the gauge! side. Unsure how that works out.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>