Remove target RSS_BYTE tracking
We previously had a mechanism that would allow the generators to back off
if the target was determined to be nearing a preconfigured memory consumption.
We decoupled this at some point -- I'm not sure when -- and then forgot to
remove the interface. It does not appear this was used in practice.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>