Remove explicit use of hyper in integration/ducks
I can't figure out how to update hyper in our project given the
set of interlinked crates that all have to update in one shot. As
a result, I'm removing explicit references to hyper and started with
ducks. This commit also removes the tower layers as it was not clear
to me that these are required for integration test purposes.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>