Upgrade hyper, tonic et al only in ducks
This commit is peer to #1179 and takes an alternative approach: only
upgrade hyper et al in single spots in the project. This means we have
a mix of different hyper versions et al but it does compile at least.

This is a wretched mess. I couldn't make tower work in ducks so it's
dropped -- we'll need to add load shedding and concurrency control ourselves
if we want it -- and while it's cool that hyper is now a 1.0 crate I
really, really don't love that we depend on a small set of 0.1 crates
to get there. I don't know if I would rather just opt into axum or not,
although hyper at least will be a going concern in the ecosystem
with high probability.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>