Generators obey throttle wait_for errors
This commit fixes a long-standing mystery: why when throughputs are low does
lading _sometimes_ emit faster than it should. It turns out the answer is
we failed to handle the capacity error that the throttle kicks out in such
conditions systematically.

This also implies that unless the user explicitly sets a max block size then
we should configure it for them, so the default of a flat 1MiB does not serve
well. I'll fix that in a follow-on PR.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>