Convert opentelemetry_metric to use MemberGenerator method
This commit adapts the OTel metrics payload internals to use the same
internal mechanism as dogstatsd: a member generator that is capable of
being more fully configured than the previous full-random approach. The
intention here is to support a similar configuration surface among both
payloads.

I have added the ability for users to customize their metric weights,
in a like manner to dogstatsd. In my follow-on commits I will add more
of the dogstatsd surface area and interior implementation, like the
use of templates to avoid reallocation of the same small strings
over and over again.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>