Generate OTel metrics with context limits
This commit allows users to configure an otel metrics payload that
has a fixed number of 'contexts' to reuse the term from dogstatsd.
This is done by setting a lot of fiddly knobs and the user is
responsible for keeping track of how many unique instances are
possible.

This commit will be chased with some clean-up as I left a lot of
internal duplication in the code. I am also unsure if this generates
sensible load run against a collector, just that it obeys the
protocol by construction.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>