Generate units according to OTel required spec
This commit adds a generator for UCUM units and replaces the purely random
strings we were using previously. We use an explicitly enumeraged table which
won't work forever but probably will work for many years until we learn
otherwise. Note that UCUM specifies a number of units do not produce.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>