Introduce RejectSet
We create and then toss a large number of blocks at startup time.
We do this because we randomly sample the state space with no rejections.
Because of the way blocks are constructed it stands to reason that there
are ranges of invalid samples and this structure is used to avoid those
invalid ranges.

Hyperfine results are now min 1.687, max 1.707 with the stdev reduced to
0.006 from 0.018. The block reject rate is now lower, which will improve
our large entropy generation non-trivially.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>