Allow lading_payload::Serialize to be mutable
This commit allows the Serialize trait defined in lading_payload
to mutate itself. This will improve the ergonimics of writing
a serializer that must have some form of memory, without requiring
other serializers to opt into this behavior.

Curiously, this did turn up two problems with dogstatsd: an aliased
function and a framing bug. Both are now corrected. I'm not sure
why this didn't trigger before.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>