All points start at 0, accumulate upward by random amount
This commit changes how points are randomly created, being zero valued. This
means that the Opentelemetry generator is responsible for setting point values
as they are generated, which is ideal. Allows us to prove that monotonic
increases happen and we reduce the computation burden on the pool.

This is effectively what we were doing already -- the underlying point value
was entirely discarded.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>