Introduce data_points_transmitted / second in OTel metrics HTTP, GRPC
This commit implements a new piece of telemetry: data points transmitted per
second. This is the equivalent of bytes per second but, when applicable, in
terms of the metric (or other) "data points" transmitted per second. This
approach should be immediately applicable to DogStatsD. This further opens the
door to other block-based metadata.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>