Create Otlp specific blackhole
The handling of Otlp is specialized enough that having a distinct
blackhole for it makes good sense, especially because we want to be
able to support gRPC request/response and could not with our existing
blackhole. In a subsequent PR I will remove the OtlpMetrics support
in the http blackhole. I believe I also understand how how to make
an otlp_grpc generator.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>