Create Otlp specific blackhole (#1365)
* Create Otlp specific blackhole

The handling of Otlp is specialized enough that having a distinct
blackhole for it makes good sense, especially because we want to be
able to support gRPC request/response and could not with our existing
blackhole. In a subsequent PR I will remove the OtlpMetrics support
in the http blackhole. I believe I also understand how how to make
an otlp_grpc generator.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* use signal labels in otlp/http

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* remove signal-specific labels

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* handle JSON encoded otlp in http endpoints

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>