Update bollard dependency
This commit is an attempt to demonstrate the claude CLI tool for
the purposes of mostly mechanical dependency upgrades. This kind
of work is a non-trivial effort sink for lading and other SMP
projects, it'd be swell if a machine could do it.

I find the result here satisfying. I acted as critical reviewer,
claude as implementor.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>