OpenTelemetry logs improvement
This commit updates our OTel logs payload to be configurable by the end user. We
allow in a manner similar to metrics for the contexts to be capped, attributes
per level of the message format to be set and seprately from the context
consideration we allow for generation and constraint of total trace-ids.

REF SMPTNG-659

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>