Allow block cache to have a 'handle' concept
With the parallel_connection addition to unix_stream we now must solve
the problem of high connection count causing non-trivial cache duplication.
This is done by allowing for a 'handle' on the cache which maintains a separate
index per handle, meaning we do not need to synchronize the cache indexes between
peers and peers are able to reference the singular cache.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>