Add a linear throttle (#1404)
### What does this PR do?

This commit allows for a linear throttle -- from min to max with a constant rate of change -- to be set by users. This is a breaking configuration change technically: it is no longer valid to set both `bytes_per_second` and an explicit throttle. Setting `bytes_per_second` implies stable throttle, as is currently true.