Introduce a 'timeout' into the stable throttle
This commit allows us to model a timeout mechanism in the stable throttle. This
is exposed as a configuration parameter on the throttle itself. The goal with
this change is to model scenarios where clients set timeout values on writes to
targets, resulting in burst load to said target when it backs up. This is
achieved by storing a ring buffer of 'unused' capacity per tick interval, with a
maximum total number of look-back, and does not require us to queue real
requests anywhere. The key insight is that a phantom request is sufficient,
meaning we can still perform blocking writes in generators and measure the
consequence of queued up requests without needing to introduce the
non-determinism of real queued up requests.

This modeling breaks down if the target actively tracks the number of queued
requests, as for instance some logs targets do. I believe Vector's end-to-end
ack is another example where this mechanism does not properly model the real
world. But, for serial requests as is the norm today in lading I think this
ought to work.

The proofs are here expanded to cover the non-zero timeout case and I have
maintained the old Valve to demonstrate that if timeout=0 the new implementation
and the old are identical.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>