Correct a bug in Opentelemetry common pool
This commit demonstrates and then corrects a bug in the template
pooling mechanism common to both metrics and logs, explaining a
mystery discovered by @ogaca-dd in PR #1447. The pool is a mechanism
to store pre-computed templates with metadata on size, allowing for
calls to OpentelemetryMetrics::to_bytes to scan for a template by byte
size, then fill in the continuously changing timestamp etc. When a
sub-call to SizedGenerator::generate failed we would _not_ search for
any template that might match, meaning subsequent calls to the pool's
generate _could_ succeed. This meant that in to_bytes I had not
previously broken the bytes_remaining >= SMALLEST_PROTOBUF loop, as
sometimes we _could_ get an appropriate template back.

We now break the loop when calls to generate fail in
OpentelemetryMetrics::to_bytes as we are guaranteed that there is no
suitable template available in the template pool.

Tests to demonstrate the problem have been added to the pool
implementation.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>