Fix long name handling in fuzzer creation (#1461)
### What does this PR do?

Let the API handle the name that are too long.
It'll return the correctly formatted name (and possibly modified) names.
The API still does need valid characters as part of the apps. This helps avoid typo (but we probably should make something more consistent at some point...)

### Motivation

Failure in Gitlab job 1086075438 (L361)


### Additional Notes

I've changed the order / grouping of operation in the try catch in order to let the other fuzzers be created if any issues is encountered on one.
Additional monitor will be added to alert us if something is going wrong, like we do in other repositories.