Add debugging logs to fuzz targets, fix OOM in otel logs fuzzers
This commit adds a FUZZ_DEBUG catch to all our lading_payload fuzz
targets to make reporting of a failing input very time cheap. This
commit also fixes an effectively unbounded trace cardinality ask in
the OTel logs fuzz targets, resulting in OOMs of the target.

I believe this situation can be improved on but I will do that in a
follow-up PR.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>