Improve DogStatsD payload generation OOMs
We are seeing that certain dogstatsd payload setups will cause OOMs
with a limit of 2GiB. This is not certain, meaning that it's
marginal. What needs to be done is to push the maximum context so far
above the current cap of 1M that we gain sufficient overhead to not
OOM.

This is done in this PR by interning strings for tags et al, avoiding
duplication of the same. We have improved memory consumption of this
payload non-trivially. The maximum number of contexts is now 10M and
generation performance as measured by `payloadtool` on
`examples/dogstatsd-generation.yaml` compared to main is improved from
mean 2.024s to 1.601s.

This interning approach is applicable to the OTel payloads as well, I
believe.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>