Improve DogStatsD payload generation OOMs (#1472)
### What does this PR do?

We are seeing that certain dogstatsd payload setups will cause OOMs with a limit of 2GiB. This is not certain, meaning that it's marginal. What needs to be done is to push the maximum context so far above the current cap of 1M that we gain sufficient overhead to not OOM.

This is done in this PR by removing an unused prefix feature on the dogstatsd generator avoiding runtime allocation of strings. We have improved memory consumption of this payload non-trivially. The maximum number of contexts is now 10M and generation performance as measured by `payloadtool` on `examples/dogstatsd-generation.yaml` compared to main is improved from mean 2.0s to 1.6s.

### Motivation

OOMs detected in SMPTNG-708