Increase testing of the K8s generator (#1482)
* Increase testing of the K8s generator

This commit increases the test surface area of the k8s generator in
two ways. Firstly, the implicit state machine is removed to an explict
module and property tested there, leaving IO encumbered code in
kubernetes.rs. In kubernetes.rs we now rely on the lading_throttle
mechanism to maintain the flow of time, avoiding coordinated omission
owing to background tokio scheduling. We further have used the mock
facility of kube-rs to make a small API to assert that the IO code
functions in the happy path as expected.

Having been confused while working on this test code I changed the
configuration to have `max_instance_lifetime_seconds` instead of
`max_lifetime_seconds` and `concurrent_instances` replaces
`number_of_instances`.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* config validation ding

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* Contain `Either` type inside resources.rs

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>