Bound OTel payloads overhead
This commit alters the fixed block cache to allow for a fixed payload
overhead. When payloads generate they need some internal allocation
space temporarily to populate a block cache. The size of this space
varies per payload but is, in the case of OTel, dependent on the
number of contexts set by the end user. Users do not have enough
information to know how much memory is required per context, that
said, and so it's very likely that a user may configure lading to
consume non-trivial memory, even if the actual amount of data they
want to compute is relatively small, as determined by the prebuilt
block cache size.

In this commit we alter the interface to the block cache, set the
maximum overhead of the payloads to the size of the prebuild cache --
downsides of this approach documented in-code -- and rig the OTel
metrics pipeline appropriately.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>