Introduce `Handle` interface to block cache, avoid duplication in HTTP generator
This commit introduces a new `Handle` API into the block cache,
inspired by the approach we took in logrotate_fs but with the goal of
being generally usuable. I have only converted the HTTP generator to
demonstrate the utility of this approach in a generator with multiple
parallel connections. My intention is to convert TCP generator after
this, then all generators in a single PR. This last PR will remove the
old non-Handle API.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>