Introduce a fingerprint mechanism for lading payloads (#1519)
* Introduce a fingerprint mechanism for lading payloads

This commit introduces a --fingerprint and --verify flag to
payloadtool. This allows us to demonstrate a sha256 for a generated
block cache, for a given config. I have also introduce a
ci/fingerprint to this project that expects configuration files in
ci/fingerprints, compared against the `fingerprint.txt` instances
stored in the same location.

For the covered configurations we are now able to determine that
payloads do not vary PR over PR unintentionally.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* clippy ding

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* run payloadtool in release mode in fingerprint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* cargo fmt ding

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* remove FxHashMap from serialization in fluent payload

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* update fingerprint

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>