Trace-agent generator
This commit introduces a new Datadog Agent traces generator. Unlike other
generators it is specific to one kind of payload -- trace_agent -- and it is
obeys backoff signal. Having implemented this it's _essentially_ the http
generator with hard-coded values so in the future we may keep the configuration
face of this generator but rely on a more generic HTTP generator.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>