Replace CaptureManager registry
This commit replaces the capture manager's internal registry with a
time-series implementation of our own devices, the goal being to
support a future need to ingest historical data. Specifically, in
order to support Datadog trace-agent we will have to support a
'dogstatsd' blackhole that is able to capture v2 series information
and forward that into the capture files. Because this is done
periodically we now have to capture historical data.

I have done this in a manner similar to logrotate_fs. There is a pure
data structure called State that is timeless and heavily property
tested that encodes all the logic we require from a time series
store. Then the wrapping code that actually implement Registry calls
into this pure structure. Likewise we maintain a way to bulk load data
with associated timestamps.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>