Add a fuzz harness for capture file validity
This commit adds an additional fuzz harness to assert that capture files
produced by lading_capture are valid. This runs the CaptureManager in afl++ in
forking mode, meaning we exercise the global registry, historical metrics et al
without needing to elide any part of the novel machinery here.

This fuzz test is however very slow. We rely on tokio's pause/advance to manage
time but it's still not quick to run. I have not found any validation issues in
produced capture files.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>