observer: prep docs to expose observer as optional
I think the quickest way to make an observer optional in Lading is to
add an enumerated type specifying the variants of observer that are
available. While the Linux observer is currently the only observer
variant available, it may be desirable to add others in the
future. Compared to a boolean option, using an enumerated type more
easily enables later extension of the observer concept.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>