observer: foreshadow making it part of lading.yaml
I'd eventually like to make the observer an optional part of Lading
enabled by adding `observer: linux` to `lading.yaml` files. In order
to do that, I need to add some types to define a Linux observer
variant and to store its variant-specific configuration. This commit
adds those types.

The `core::Default` implementations written in this commit are
intended to be temporary to preserve the existing `lading.yaml`
semantics that treat the observer as mandatory when `lading` is run
with a target.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>