Introduce configuration directory support
This commit adds support to lading for configuration directories. All
files in the directory are required to be disjoint. If they are not,
lading will fail with an error. This change supports "system" configs
and "user" configs where one may have a user-configurable set of
configuration files and a static system config.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>