Improve log message in containers generator, correct deletion bug
This commit expands the error messages of the containers generator to
have more contextual information. Likewise, we allow for some former
error conditions -- like a 404 from bollard -- to count as
success. That is, formerly when we attempted to clean a container _if_
that container had been otherwise removed lading would issue an
error. However, lading's goal was to see the container removed, not
strictly to be the one to remove it.

Containers organized in pods and monitored externally to lading are
prone to this behavior manifesting.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>