Drive proptest configuration via envars
Our ci/validate script runs for multiple minutes locally. This hampers
local development iterations. However, the property tests that run for
this duration do catch bugs. So, we increase the iterations in CI and
decrease them in local development, transitioning the project to use
one global setup for all property tests instead of a test by test
approach as previously taken.

It's unclear as of this commit if the CI settings are too aggressive.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>